third-python-layer/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
google-generativeai==0.3.2
elasticsearch==8.11.1
python-multipart==0.0.6
pydantic==2.6.1
python-dotenv==1.0.1
PyPDF2==3.0.1
numpy==1.26.3
sentence-transformers==2.2.2
pillow==10.2.0 